# Improve performance of expression evaluation