--- a/cpp/include/cudf/ast/detail/expression_evaluator.cuh
+++ b/cpp/include/cudf/ast/detail/expression_evaluator.cuh
@@ -317,7 +317,7 @@ struct expression_evaluator {
       } else {
         return ReturnType(plan.literals[input_reference.data_index].value<Element>());
       }
-    } else {  // Assumes input_reference.data_index ==
+    } else {  // Assumes input_reference.reference_type ==
               // detail::device_data_reference_type::INTERMEDIATE
       // Using memcpy instead of reinterpret_cast<Element*> for safe type aliasing
       // Using a temporary variable ensures that the compiler knows the result is aligned
@@ -550,7 +550,7 @@ struct expression_evaluator {
     {
       if (device_data_reference.reference_type == detail::device_data_reference_type::COLUMN) {
         output_object.template set_value<Element>(row_index, result);
-      } else {  // Assumes input_reference.data_index ==
+      } else {  // Assumes device_data_reference.reference_type ==
                 // detail::device_data_reference_type::INTERMEDIATE
         // Using memcpy instead of reinterpret_cast<Element*> for safe type aliasing.
         // Using a temporary variable ensures that the compiler knows the result is aligned.