@vyasr 2021-09-09
Here are some benchmarks

Conditional joins on a Tesla T4:
<details><summary>Benchmarks</summary>
<p>

Before:
```
----------------------------------------------------------------------------------------------------------------------------------------------
Benchmark                                                                                                    Time             CPU   Iterations
----------------------------------------------------------------------------------------------------------------------------------------------
ConditionalJoin<int32_t, int32_t>/conditional_inner_join_32bit/100000/100000/manual_time                   996 ms          996 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_inner_join_32bit/100000/400000/manual_time                  3368 ms         3368 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_inner_join_32bit/100000/1000000/manual_time                 8386 ms         8385 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_inner_join_64bit/100000/100000/manual_time                   877 ms          877 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_inner_join_64bit/100000/400000/manual_time                  3505 ms         3505 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_inner_join_64bit/100000/1000000/manual_time                 8713 ms         8713 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_inner_join_32bit_nulls/100000/100000/manual_time            1591 ms         1591 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_inner_join_32bit_nulls/100000/400000/manual_time            6008 ms         6008 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_inner_join_32bit_nulls/100000/1000000/manual_time          14905 ms        14904 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_inner_join_64bit_nulls/100000/100000/manual_time            1644 ms         1644 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_inner_join_64bit_nulls/100000/400000/manual_time            6241 ms         6241 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_inner_join_64bit_nulls/100000/1000000/manual_time          15555 ms        15554 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_left_join_32bit/100000/100000/manual_time                    876 ms          876 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_left_join_32bit/100000/400000/manual_time                   3503 ms         3503 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_left_join_32bit/100000/1000000/manual_time                  8662 ms         8662 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_left_join_64bit/100000/100000/manual_time                    901 ms          901 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_left_join_64bit/100000/400000/manual_time                   3604 ms         3604 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_left_join_64bit/100000/1000000/manual_time                  8928 ms         8927 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_left_join_32bit_nulls/100000/100000/manual_time             1621 ms         1621 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_left_join_32bit_nulls/100000/400000/manual_time             6149 ms         6148 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_left_join_32bit_nulls/100000/1000000/manual_time           15182 ms        15182 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_left_join_64bit_nulls/100000/100000/manual_time             1654 ms         1654 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_left_join_64bit_nulls/100000/400000/manual_time             6303 ms         6303 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_left_join_64bit_nulls/100000/1000000/manual_time           15621 ms        15621 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_full_join_32bit/100000/100000/manual_time                    877 ms          877 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_full_join_32bit/100000/400000/manual_time                   3510 ms         3510 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_full_join_32bit/100000/1000000/manual_time                  8661 ms         8661 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_full_join_64bit/100000/100000/manual_time                    903 ms          903 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_full_join_64bit/100000/400000/manual_time                   3599 ms         3599 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_full_join_64bit/100000/1000000/manual_time                  8900 ms         8900 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_full_join_32bit_nulls/100000/100000/manual_time             1617 ms         1617 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_full_join_32bit_nulls/100000/400000/manual_time             6116 ms         6115 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_full_join_32bit_nulls/100000/1000000/manual_time           15083 ms        15082 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_full_join_64bit_nulls/100000/100000/manual_time             1648 ms         1648 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_full_join_64bit_nulls/100000/400000/manual_time             6269 ms         6269 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_full_join_64bit_nulls/100000/1000000/manual_time           15527 ms        15526 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_left_anti_join_32bit/100000/100000/manual_time               872 ms          872 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_left_anti_join_32bit/100000/400000/manual_time              3484 ms         3484 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_left_anti_join_32bit/100000/1000000/manual_time             8612 ms         8611 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_left_anti_join_64bit/100000/100000/manual_time               897 ms          897 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_left_anti_join_64bit/100000/400000/manual_time              3579 ms         3579 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_left_anti_join_64bit/100000/1000000/manual_time             8857 ms         8856 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_left_anti_join_32bit_nulls/100000/100000/manual_time        1617 ms         1617 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_left_anti_join_32bit_nulls/100000/400000/manual_time        6077 ms         6076 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_left_anti_join_32bit_nulls/100000/1000000/manual_time      14981 ms        14980 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_left_anti_join_64bit_nulls/100000/100000/manual_time        1630 ms         1630 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_left_anti_join_64bit_nulls/100000/400000/manual_time        6250 ms         6250 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_left_anti_join_64bit_nulls/100000/1000000/manual_time      15428 ms        15427 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_left_semi_join_32bit/100000/100000/manual_time               869 ms          869 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_left_semi_join_32bit/100000/400000/manual_time              3475 ms         3475 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_left_semi_join_32bit/100000/1000000/manual_time             8593 ms         8592 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_left_semi_join_64bit/100000/100000/manual_time               895 ms          895 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_left_semi_join_64bit/100000/400000/manual_time              3567 ms         3567 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_left_semi_join_64bit/100000/1000000/manual_time             8830 ms         8830 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_left_semi_join_32bit_nulls/100000/100000/manual_time        1601 ms         1601 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_left_semi_join_32bit_nulls/100000/400000/manual_time        6063 ms         6063 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_left_semi_join_32bit_nulls/100000/1000000/manual_time      14941 ms        14941 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_left_semi_join_64bit_nulls/100000/100000/manual_time        1640 ms         1640 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_left_semi_join_64bit_nulls/100000/400000/manual_time        6220 ms         6220 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_left_semi_join_64bit_nulls/100000/1000000/manual_time      15418 ms        15417 ms            1
```

After:

```
----------------------------------------------------------------------------------------------------------------------------------------------
Benchmark                                                                                                    Time             CPU   Iterations
----------------------------------------------------------------------------------------------------------------------------------------------
ConditionalJoin<int32_t, int32_t>/conditional_inner_join_32bit/100000/100000/manual_time                   818 ms          818 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_inner_join_32bit/100000/400000/manual_time                  2465 ms         2465 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_inner_join_32bit/100000/1000000/manual_time                 6088 ms         6088 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_inner_join_64bit/100000/100000/manual_time                   677 ms          677 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_inner_join_64bit/100000/400000/manual_time                  2541 ms         2541 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_inner_join_64bit/100000/1000000/manual_time                 6305 ms         6305 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_inner_join_32bit_nulls/100000/100000/manual_time            1470 ms         1470 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_inner_join_32bit_nulls/100000/400000/manual_time            5584 ms         5584 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_inner_join_32bit_nulls/100000/1000000/manual_time          13852 ms        13851 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_inner_join_64bit_nulls/100000/100000/manual_time            1601 ms         1601 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_inner_join_64bit_nulls/100000/400000/manual_time            6034 ms         6033 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_inner_join_64bit_nulls/100000/1000000/manual_time          15085 ms        15085 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_left_join_32bit/100000/100000/manual_time                    690 ms          690 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_left_join_32bit/100000/400000/manual_time                   2607 ms         2607 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_left_join_32bit/100000/1000000/manual_time                  6431 ms         6431 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_left_join_64bit/100000/100000/manual_time                    709 ms          710 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_left_join_64bit/100000/400000/manual_time                   2679 ms         2679 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_left_join_64bit/100000/1000000/manual_time                  6629 ms         6629 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_left_join_32bit_nulls/100000/100000/manual_time             1563 ms         1563 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_left_join_32bit_nulls/100000/400000/manual_time             5856 ms         5856 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_left_join_32bit_nulls/100000/1000000/manual_time           14455 ms        14454 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_left_join_64bit_nulls/100000/100000/manual_time             1655 ms         1655 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_left_join_64bit_nulls/100000/400000/manual_time             6258 ms         6257 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_left_join_64bit_nulls/100000/1000000/manual_time           15513 ms        15512 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_full_join_32bit/100000/100000/manual_time                    705 ms          705 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_full_join_32bit/100000/400000/manual_time                   2687 ms         2687 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_full_join_32bit/100000/1000000/manual_time                  6602 ms         6602 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_full_join_64bit/100000/100000/manual_time                    723 ms          723 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_full_join_64bit/100000/400000/manual_time                   2746 ms         2746 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_full_join_64bit/100000/1000000/manual_time                  6779 ms         6779 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_full_join_32bit_nulls/100000/100000/manual_time             1579 ms         1579 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_full_join_32bit_nulls/100000/400000/manual_time             5968 ms         5967 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_full_join_32bit_nulls/100000/1000000/manual_time           14658 ms        14657 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_full_join_64bit_nulls/100000/100000/manual_time             1670 ms         1670 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_full_join_64bit_nulls/100000/400000/manual_time             6338 ms         6338 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_full_join_64bit_nulls/100000/1000000/manual_time           15621 ms        15621 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_left_anti_join_32bit/100000/100000/manual_time               708 ms          708 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_left_anti_join_32bit/100000/400000/manual_time              2691 ms         2691 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_left_anti_join_32bit/100000/1000000/manual_time             6623 ms         6623 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_left_anti_join_64bit/100000/100000/manual_time               725 ms          725 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_left_anti_join_64bit/100000/400000/manual_time              2752 ms         2752 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_left_anti_join_64bit/100000/1000000/manual_time             6781 ms         6781 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_left_anti_join_32bit_nulls/100000/100000/manual_time        1575 ms         1575 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_left_anti_join_32bit_nulls/100000/400000/manual_time        5944 ms         5944 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_left_anti_join_32bit_nulls/100000/1000000/manual_time      14639 ms        14638 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_left_anti_join_64bit_nulls/100000/100000/manual_time        1667 ms         1667 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_left_anti_join_64bit_nulls/100000/400000/manual_time        6322 ms         6322 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_left_anti_join_64bit_nulls/100000/1000000/manual_time      15561 ms        15558 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_left_semi_join_32bit/100000/100000/manual_time               708 ms          708 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_left_semi_join_32bit/100000/400000/manual_time              2683 ms         2683 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_left_semi_join_32bit/100000/1000000/manual_time             6596 ms         6595 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_left_semi_join_64bit/100000/100000/manual_time               722 ms          722 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_left_semi_join_64bit/100000/400000/manual_time              2739 ms         2739 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_left_semi_join_64bit/100000/1000000/manual_time             6755 ms         6755 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_left_semi_join_32bit_nulls/100000/100000/manual_time        1569 ms         1569 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_left_semi_join_32bit_nulls/100000/400000/manual_time        5944 ms         5943 ms            1
ConditionalJoin<int32_t, int32_t>/conditional_left_semi_join_32bit_nulls/100000/1000000/manual_time      14550 ms        14550 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_left_semi_join_64bit_nulls/100000/100000/manual_time        1659 ms         1659 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_left_semi_join_64bit_nulls/100000/400000/manual_time        6273 ms         6273 ms            1
ConditionalJoin<int64_t, int64_t>/conditional_left_semi_join_64bit_nulls/100000/1000000/manual_time      15508 ms        15508 ms            1
```
</p>
</details>

`compute_column` on a Tesla T4:
<details><summary>Benchmarks</summary>
<p>

Before:
```
--------------------------------------------------------------------------------------------------------------------------------------------------------------------------
Benchmark                                                                                                                Time             CPU   Iterations UserCounters...
--------------------------------------------------------------------------------------------------------------------------------------------------------------------------
AST<int32_t, TreeType::IMBALANCED_LEFT, false, false>/ast_int32_imbalanced_unique/100000/1/manual_time               0.024 ms        0.048 ms        29405 bytes_per_second=31.6867G/s
AST<int32_t, TreeType::IMBALANCED_LEFT, false, false>/ast_int32_imbalanced_unique/100000/5/manual_time               0.038 ms        0.063 ms        18501 bytes_per_second=59.0205G/s
AST<int32_t, TreeType::IMBALANCED_LEFT, false, false>/ast_int32_imbalanced_unique/100000/10/manual_time              0.057 ms        0.076 ms        12347 bytes_per_second=72.29G/s
AST<int32_t, TreeType::IMBALANCED_LEFT, false, false>/ast_int32_imbalanced_unique/1000000/1/manual_time              0.073 ms        0.094 ms         9631 bytes_per_second=102.042G/s
AST<int32_t, TreeType::IMBALANCED_LEFT, false, false>/ast_int32_imbalanced_unique/1000000/5/manual_time              0.211 ms        0.231 ms         3355 bytes_per_second=106.132G/s
…